HDDS-9238. Moved web.utils methods to s3g
