+++ hadoop-ozone/s3gateway/src/main/java/org/apache/hadoop/ozone/s3/RequestIdentifier.java
@@ -19,12 +19,12 @@ package org.apache.hadoop.ozone.s3;
 
 import javax.enterprise.context.RequestScoped;
 
-import org.apache.hadoop.ozone.web.utils.OzoneUtils;
-
 import org.apache.commons.lang3.RandomStringUtils;
 
 import java.security.SecureRandom;
 import java.util.Random;
+
+import static org.apache.hadoop.ozone.s3.util.S3Utils.getRequestID;
 
 /**
  * Request specific identifiers.
@@ -41,7 +41,7 @@ public class RequestIdentifier {
   public RequestIdentifier() {
     int count = 8 + RANDOM.nextInt(8);
     amzId = RandomStringUtils.random(count, 0, 0, true, true, null, RANDOM);
-    requestId = OzoneUtils.getRequestID();
+    requestId = getRequestID();
   }
 
   public String getRequestId() {
