+++ hadoop-ozone/s3gateway/src/test/java/org/apache/hadoop/ozone/s3/exception/TestOS3Exceptions.java
@@ -18,9 +18,10 @@
 
 package org.apache.hadoop.ozone.s3.exception;
 
-import org.apache.hadoop.ozone.web.utils.OzoneUtils;
 import org.junit.Assert;
 import org.junit.Test;
+
+import static org.apache.hadoop.ozone.s3.util.S3Utils.getRequestID;
 
 /**
  * This class tests OS3Exception class.
@@ -31,7 +32,7 @@ public class TestOS3Exceptions {
   public void testOS3Exceptions() {
     OS3Exception ex = new OS3Exception("AccessDenied", "Access Denied",
         403);
-    String requestId = OzoneUtils.getRequestID();
+    String requestId = getRequestID();
     ex = S3ErrorTable.newError(ex, "bucket");
     ex.setRequestId(requestId);
     String val = ex.toXml();
